+++ processor/src/operations/mod.rs
@@ -145,7 +145,7 @@ where
             // ----- cryptographic operations -----------------------------------------------------
             Operation::HPerm => self.op_hperm()?,
             Operation::MpVerify => self.op_mpverify()?,
-            Operation::MrUpdate(copy) => self.op_mrupdate(copy)?,
+            Operation::MrUpdate => self.op_mrupdate()?,
             Operation::FriE2F4 => self.op_fri_ext2fold4()?,
         }